Move some code into its own files
